+++ src/Api/Request/GetOrder.php
@@ -17,7 +17,7 @@ class GetOrder extends Request
     public function __construct(string $orderId)
     {
         $this->setRequestMethod('GET');
-        $this->setApiEndpointPath(sprintf('orders/%s', $orderId));
+        $this->setApiEndpointPath(sprintf('orders/%s', rawurlencode($orderId)));
     }
 
     /**